requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.9
django-cors-headers==4.9.0
django-extensions==3.2.3
django-health-check==3.18.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
gunicorn==23.0.0
httplib2==0.31.1
idna==3.11
oauthlib==3.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.3.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.4
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.9.0
